refactor: move `filetest` internal package out of `fileconsumer` package
To reuse it in File input benchmark.
