+++ pkg/stanza/internal/filetest/filetest.go
