+++ pkg/stanza/internal/filetest/package_test.go
